refactor(storage): extract Phase 3a dual_write helpers (R.5)
Phase R.5 of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Moves the entire Phase 3a
dual-write subsystem out of store.rs (5599 → 4975 lines) into a new
`dual_write.rs` (653 LOC). Bonus: this whole file gets DELETED in
Phase 3c when the legacy `db_agent_definitions` / `db_agent_instances`
tables are dropped — isolating it now makes that future deletion a
clean unit.

Methods (pub(crate)):
- `agents_dual_write_definition_upsert`
- `agents_dual_write_definition_delete`
- `agents_dual_write_seeded_delete`
- `agents_dual_write_instance_create`
- `agents_dual_write_instance_update`
- `agents_dual_write_instance_set_hidden`
- `agents_dual_write_instance_repoint`
- `agents_dual_write_instance_delete`
- `agents_dual_write_backfill_identity`

Private helpers (associated functions on Store):
- `agents_projection_key_for_inst` (CTE chain-walk for projection key)
- `load_definition_for_dual_write` (in-lock def re-read)

Method surface is unchanged — same `impl Store {}` block pattern.
No external imports affected: all dual_write methods are `pub(crate)`,
and the helpers are private associated functions only called from
within the dual_write module.

Net store.rs change: -624 lines. New file 653 LOC.

162/162 storage tests green.
